feat: add confirmation before deleting job

+++ src/views/JobView.vue
@@ -19,9 +19,13 @@ const state = reactive({
 
 const deleteJob = async ()=>{
   try {
-   await axios.delete(`${import.meta.env.VITE_BASE_URL}/jobs/${jobId}`) 
-   toast.success("Job deleted successfully")
-   router.push('/jobs')
+    const confirm = window.confirm('Are you sure you want to delete this job?')
+    if(confirm){
+
+      await axios.delete(`${import.meta.env.VITE_BASE_URL}/jobs/${jobId}`) 
+      toast.success("Job deleted successfully")
+      router.push('/jobs')
+    }
   } catch (error) {
     console.error('error deleting job', error)
     toast.error('Error deleting job')
